feat: Implement fuzzy matching for search optimization
Previously find function only finds tasks that exactly matches with user
input. This is not user-friendly as the user might remember the
description wrongly or simply has typo error.

A fuzzy matching algorithm (Levenstein in this case) is needed to
accommodate such type errors and still find the tasks the user might be
asking for.